requirements.txt
pillow
datasets
sentence_transformers
faiss-cpu